README
This Repository is meant to express my interest in video games. Now, the code I have worked on through a book I have received is no where near the games you see in todays day, but rather games like a number guessing game, or a dragon exploration decision type game. This Repository will be filled with these kinds of games as I progress through the book.